+++ packages/server/src/common/command/command-stack.ts
@@ -13,62 +13,131 @@
  *
  * SPDX-License-Identifier: EPL-2.0 OR GPL-2.0 WITH Classpath-exception-2.0
  ********************************************************************************/
-import { injectable } from 'inversify';
+import { inject, injectable } from 'inversify';
+import { Logger } from '../utils/logger';
 import { Command } from './command';
 
 export const CommandStack = Symbol('CommandStack');
 
 /**
- * A basic implementation of an undoable stack of commands.
- * See {@link Command} for more details about the command methods that this implementation uses.
+ * A command stack keeps track of {@link Command commands} that have been executed.
+ * Executed commands can be undone in the reverse order of their original execution.
+ * Unless a new command has been executed after an undo operation, commands can be
+ * redone in the original order of their execution.
+ *
  */
 export interface CommandStack {
     /**
-     * Clears any redoable commands not yet redone, adds the command, and then executes the command.
+     * Clears any redoable commands not yet redone, adds the command to the stack and then invokes {@link Command.execute}.
      * @param command the command to execute.
      */
     execute(command: Command): void;
 
     /**
-     * Moves the top of the stack down, undoing what was formerly the top command.
+     * Removes the topmost (i.e. last executed) command from the stack and invokes {@link Command.undo}.
      */
     undo(): void;
 
     /**
-     * Moves the top of the stack up, redoing the new top command.
+     * Returns `true` if the top command on the stack can be undone.
+     */
+    canUndo(): boolean;
+
+    /**
+     * Re-adds the last undo command on top of the stack and invokes {@link Command.redo}
      */
     redo(): void;
 
     /**
-     * The command stack has currently unsaved changes.
+     * Returns `true` if there are  redoable commands in the stack.
      */
-    isDirty(): boolean;
+    canRedo(): boolean;
 
     /**
      * Called after a save has been successfully performed.
      */
     saveIsDone(): void;
+
+    /**
+     * Indicates wether the command stack has unsaved changes.
+     */
+    readonly isDirty: boolean;
+
+    /**
+     * Disposes all the commands in the stack.
+     */
+    flush(): void;
 }
 
 @injectable()
 export class DefaultCommandStack implements CommandStack {
+    @inject(Logger)
+    protected logger: Logger;
+
+    protected undoStack: Command[] = [];
+    protected redoStack: Command[] = [];
+
     execute(command: Command): void {
-        // no-op
+        this.redoStack = [];
+        try {
+            command.execute();
+        } catch (error) {
+            this.handleError(error);
+        }
+        this.undoStack.push(command);
     }
 
     undo(): void {
-        // no-op
+        const command = this.undoStack.pop();
+        if (command && Command.canUndo(command)) {
+            try {
+                command.undo();
+            } catch (error) {
+                this.handleError(error);
+            }
+            this.redoStack.push(command);
+        }
+    }
+
+    canUndo(): boolean {
+        return this.undoStack.length > 0 //
+            ? Command.canUndo(this.undoStack[this.undoStack.length - 1])
+            : false;
     }
 
     redo(): void {
-        // no-op
+        const command = this.redoStack.pop();
+        if (command) {
+            try {
+                command.redo();
+            } catch (error) {
+                this.handleError(error);
+            }
+            this.undoStack.push(command);
+        }
     }
 
-    isDirty(): boolean {
-        return false;
+    canRedo(): boolean {
+        return this.redoStack.length > 0;
+    }
+
+    get isDirty(): boolean {
+        return this.undoStack.length > 0;
     }
 
     saveIsDone(): void {
-        // no-op
+        this.flush();
+    }
+
+    flush(): void {
+        this.undoStack = [];
+        this.redoStack = [];
+    }
+
+    protected handleError(error: any): never {
+        // if an error occurred during the command execution the stack might be in an erroneous state => we have to flush the command stack
+        this.flush();
+        this.logger.error('An error occurred during command execution. CommandStack will be flushed!', error);
+        throw error;
     }
 }