+++ packages/server/src/common/operations/operation-handler.ts
@@ -14,38 +14,63 @@
  * SPDX-License-Identifier: EPL-2.0 OR GPL-2.0 WITH Classpath-exception-2.0
  ********************************************************************************/
 import { MaybePromise, Operation } from '@eclipse-glsp/protocol';
-import { interfaces } from 'inversify';
-
-export const OperationHandler = Symbol('OperationHandler');
+import { inject, injectable, interfaces } from 'inversify';
+import { Command } from '../command/command';
+import { ModelState } from '../features/model/model-state';
 
 /**
  * An operation handler can execute {@link Operation}s of a certain type (subclass).
  * The operation handler processes the operation in the {@link OperationHandler.execute()} method. The result
- * of the execution is an update of the `ModelState` state.
- * This update is reversible (undo) and can be reapplied (redo). For basic diagram languages these updates are typically
- * applied directly on the `ModelState` using EMF `Command`s and the
- * `ModelState.execute()` method. For more complex diagram languages the
- * GModel state might be updated indirectly and the operation handler manipulates a custom model representation.
+ * of the execution is a {@link Command} that captures the corresponding source model changes.
+ * This command can be applied on the `CommandStack` and is reversible (undo) and can be reapplied (redo).
  *
  * The `OperationActionHandler` is responsible for retrieving all available (valid) operation handlers for an
  * operation that is dispatched via `ActionDispatcher`.
  */
-export interface OperationHandler {
+@injectable()
+export abstract class OperationHandler {
+    @inject(ModelState)
+    protected modelState: ModelState;
+
     /**
      * Returns the operation type that can be handled by this operation handler.
      *
      * @returns the operation type that can be handled.
      */
-    readonly operationType: string;
+    abstract operationType: string;
 
     readonly label?: string;
 
     /**
-     * Executes the operation handler for the given {@link Operation}.
+     * Creates a command that performs the operation in the source model(s). If `undefined` is  returned, no update
+     * is performed on the model(s).
      *
-     * @param operation The operation that should be executed.
+     * @param operation The operation to process.
+     * @return The created command to be executed on the command stack or `undefined` nothing should be done.
      */
-    execute(operation: Operation): MaybePromise<void>;
+    abstract createCommand(operation: Operation): MaybePromise<Command | undefined>;
+
+    /**
+     * Executes the operation handler for the given {@link Operation} and returns the corresponding {@link Command}.
+     * If the given operation cannot be handled by this handler or the handler execution did not result in any changes
+     * `undefined` is returned.
+     *
+     * @param operation The operation that should be executed or empty if nothing should be done.
+     * @returns The command capturing the execution changes or `undefined`
+     */
+    execute(operation: Operation): MaybePromise<Command | undefined> {
+        return this.handles(operation) ? this.createCommand(operation) : undefined;
+    }
+
+    /**
+     * Validates whether the given {@link Operation} can be handled by this operation handler.
+     *
+     * @param operation The operation that should be validated.
+     * @returns `true` if the given operation can be handled, `false` otherwise.
+     */
+    handles(operation: Operation): boolean {
+        return this.modelState.root && operation.kind === this.operationType;
+    }
 }
 
 export const OperationHandlerConstructor = Symbol('OperationHandlerConstructor');